requirements.txt
python-dotenv
pandas
numpy
certifi
pymongo[srv]==4.6.3
pymongo
scikit-learn
dill
pyaml
mlflow
#-e . ##This will refer it to setup.py  we use it at the end when we want to complete or deploy our project

